requirements.txt
uvicorn==0.13.3
fastapi==0.63.0
starlette==0.13.6
pydantic==1.7.3
pydantic[email]==1.7.3
typing==3.7.4.3
mangum==0.10.0
SQLAlchemy==1.3.22
pymysql==1.0.2
psycopg2-binary==2.8.4
databases==0.4.1
asyncpg==0.21.0
python-dotenv==0.15.0
shapely==1.7.1
boto3==1.17.22
python-jose==3.1.0